requirements.txt
pandas~=1.3.5
numpy~=1.21.6
plotly~=5.13.0
easydict~=1.10
tensorboard
python-dotenv
nbformat